# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "todolist-app API",
    "description": "개인 할일 관리 웹 애플리케이션 REST API\n\n## 인증 방식\n- **Access Token**: `Authorization: Bearer <token>` 헤더 (유효기간 15분)\n- **Refresh Token**: HTTP-only 쿠키 자동 전송 (유효기간 7일)\n\n## 할일 상태(status) 계산\nstatus는 DB에 저장하지 않고 서버에서 런타임 계산:\n- `NOT_STARTED`: 오늘 < start_date\n- `IN_PROGRESS`: start_date ≤ 오늘 ≤ due_date, 미완료\n- `OVERDUE`: 오늘 > due_date, 미완료\n- `COMPLETED_SUCCESS`: is_completed=true, is_success=true\n- `COMPLETED_FAILURE`: is_completed=true, is_success=false",
    "version": "1.0.0",
    "contact": {
      "name": "Yongwoo"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "로컬 개발 서버"
    }
  ],
...
  "paths": {
...
    "/todos": {
      "post": {
        "tags": ["todos"],
        "summary": "할일 생성",
        "description": "새로운 할일을 생성합니다. (UC-03)\n\n- `due_date`는 `start_date` 이상이어야 합니다\n- `is_completed`는 기본값 `false`로 설정됩니다\n- `status`는 서버에서 `start_date`, `due_date`, 오늘 날짜(KST)를 기반으로 런타임 계산됩니다",
        "operationId": "createTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoRequest"
              },
              "example": {
                "title": "수학 수행평가 제출",
                "description": "교과서 5~7단원 문제 풀이 포함",
                "start_date": "2026-04-01",
                "due_date": "2026-04-07"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "할일 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                },
                "example": {
                  "success": true,
                  "todo": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "user_id": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "수학 수행평가 제출",
                    "description": "교과서 5~7단원 문제 풀이 포함",
                    "start_date": "2026-04-01",
                    "due_date": "2026-04-07",
                    "is_completed": false,
                    "is_success": null,
                    "status": "IN_PROGRESS",
                    "created_at": "2026-04-01T09:00:00.000Z",
                    "updated_at": "2026-04-01T09:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 검증 실패 (제목 누락, 제목 길이 초과, 날짜 형식 오류, 필수 필드 누락 등)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "제목은 1자 이상 100자 이하여야 합니다"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "description": "처리 불가 엔티티 (due_date가 start_date보다 이전, BR-03)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "종료일은 시작일과 같거나 이후 날짜여야 합니다"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "get": {
        "tags": ["todos"],
        "summary": "할일 목록 조회",
        "description": "로그인한 사용자의 할일 목록을 조회합니다. (UC-04)\n\n- 본인의 할일만 반환됩니다\n- `status` 필터 생략 시 전체 상태 반환\n- `status`는 서버에서 런타임 계산 후 포함됩니다",
        "operationId": "getTodos",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TodoStatus"
            },
            "description": "할일 상태 필터 (생략 시 전체 조회)"
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["start_date", "due_date"],
              "default": "due_date"
            },
            "description": "정렬 기준 컬럼"
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "asc"
            },
            "description": "정렬 방향"
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "페이지 번호 (1 이상)"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "페이지당 항목 수 (1~100)"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListResponse"
                },
                "example": {
                  "success": true,
                  "todos": [
                    {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "user_id": "550e8400-e29b-41d4-a716-446655440000",
                      "title": "수학 수행평가 제출",
                      "description": "교과서 5~7단원 문제 풀이 포함",
                      "start_date": "2026-04-01",
                      "due_date": "2026-04-07",
                      "is_completed": false,
                      "is_success": null,
                      "status": "IN_PROGRESS",
                      "created_at": "2026-04-01T09:00:00.000Z",
                      "updated_at": "2026-04-01T09:00:00.000Z"
                    }
                  ],
                  "total": 1,
                  "page": 1,
                  "limit": 10
                }
              }
            }
          },
          "400": {
            "description": "쿼리 파라미터 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "tags": ["todos"],
        "summary": "할일 상세 조회",
        "description": "특정 할일의 상세 정보를 조회합니다. (EP-06)\n\n본인의 할일만 조회 가능합니다.",
        "operationId": "getTodoById",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 상세 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                },
                "example": {
                  "success": true,
                  "todo": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "user_id": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "분기 보고서 작성",
                    "description": "Q1 실적 분석 포함",
                    "start_date": "2026-04-01",
                    "due_date": "2026-04-06",
                    "is_completed": false,
                    "is_success": null,
                    "status": "IN_PROGRESS",
                    "created_at": "2026-04-01T09:00:00.000Z",
                    "updated_at": "2026-04-01T09:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["todos"],
        "summary": "할일 수정",
        "description": "할일의 제목, 설명, 날짜를 수정합니다. (UC-05)\n\n- 최소 하나의 필드가 포함되어야 합니다\n- 날짜 변경 시 `due_date >= start_date` 조건 재검증\n- `updated_at`은 자동 갱신됩니다\n- 본인의 할일만 수정 가능합니다",
        "operationId": "updateTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              },
              "example": {
                "due_date": "2026-04-06"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "할일 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                },
                "example": {
                  "success": true,
                  "todo": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "user_id": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "분기 보고서 작성",
                    "description": "Q1 실적 분석 포함",
                    "start_date": "2026-04-01",
                    "due_date": "2026-04-06",
                    "is_completed": false,
                    "is_success": null,
                    "status": "IN_PROGRESS",
                    "created_at": "2026-04-01T09:00:00.000Z",
                    "updated_at": "2026-04-01T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 검증 실패 또는 수정할 필드 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "description": "처리 불가 엔티티 (due_date가 start_date보다 이전, BR-03)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "종료일은 시작일과 같거나 이후 날짜여야 합니다"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["todos"],
        "summary": "할일 삭제",
        "description": "할일을 삭제합니다. (UC-07)\n\n- 삭제된 할일은 복구 불가합니다\n- 본인의 할일만 삭제 가능합니다",
        "operationId": "deleteTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "204": {
            "description": "할일 삭제 성공 (응답 본문 없음)"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/todos/{id}/complete": {
      "post": {
        "tags": ["todos"],
        "summary": "할일 완료 처리",
        "description": "할일을 완료 상태로 전환하고 성공/실패를 기록합니다. (UC-06)\n\n- `OVERDUE` 상태의 할일도 완료 처리 가능합니다 (BR-09)\n- 이미 완료된 할일은 재완료 불가합니다\n- 본인의 할일만 완료 처리 가능합니다",
        "operationId": "completeTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTodoRequest"
              },
              "example": {
                "is_success": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "완료 처리 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteTodoResponse"
                },
                "examples": {
                  "completed_success": {
                    "summary": "성공 완료",
                    "value": {
                      "success": true,
                      "todo": {
                        "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                        "title": "분기 보고서 작성",
                        "is_completed": true,
                        "is_success": true,
                        "status": "COMPLETED_SUCCESS",
                        "updated_at": "2026-04-01T18:00:00.000Z"
                      }
                    }
                  },
                  "completed_failure": {
                    "summary": "실패 완료 (OVERDUE 경로)",
                    "value": {
                      "success": true,
                      "todo": {
                        "id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                        "title": "민원 처리 보고서 제출",
                        "is_completed": true,
                        "is_success": false,
                        "status": "COMPLETED_FAILURE",
                        "updated_at": "2026-04-01T18:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "is_success 필드 누락 또는 유효하지 않은 값",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "이미 완료된 할일",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "이미 완료 처리된 할일입니다"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access Token (유효기간 15분, HS256 알고리즘)\n\n페이로드: `{ sub: user_id, email, iat, exp }`"
      }
    },
    "parameters": {
      "TodoId": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "할일 고유 식별자 (UUID)"
      }
    },
    "schemas": {
      "TodoStatus": {
        "type": "string",
        "enum": [
          "NOT_STARTED",
          "IN_PROGRESS",
          "OVERDUE",
          "COMPLETED_SUCCESS",
          "COMPLETED_FAILURE"
        ],
        "description": "할일 상태 (서버 런타임 계산값)\n- `NOT_STARTED`: 오늘 < start_date\n- `IN_PROGRESS`: start_date ≤ 오늘 ≤ due_date, 미완료\n- `OVERDUE`: 오늘 > due_date, 미완료\n- `COMPLETED_SUCCESS`: is_completed=true, is_success=true\n- `COMPLETED_FAILURE`: is_completed=true, is_success=false"
      },
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "할일 고유 식별자"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "소유 사용자 ID"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "할일 제목"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true,
            "description": "할일 상세 내용 (선택)"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "시작일 (YYYY-MM-DD)"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "description": "종료일 (YYYY-MM-DD, start_date 이상)"
          },
          "is_completed": {
            "type": "boolean",
            "description": "사용자 완료 처리 여부"
          },
          "is_success": {
            "type": "boolean",
            "nullable": true,
            "description": "성공 여부 (is_completed=true일 때만 유효, 미완료 시 null)"
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "생성 일시 (ISO 8601)"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "최종 수정 일시 (ISO 8601)"
          }
        },
        "required": [
          "id", "user_id", "title", "start_date", "due_date",
          "is_completed", "status", "created_at", "updated_at"
        ]
      },
...
      "CreateTodoRequest": {
        "type": "object",
        "required": ["title", "start_date", "due_date"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "할일 제목 (1~100자)"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "description": "할일 상세 내용 (선택, 최대 1000자)"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "시작일 (YYYY-MM-DD)"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "description": "종료일 (YYYY-MM-DD, start_date 이상)"
          }
        }
      },
      "UpdateTodoRequest": {
        "type": "object",
        "description": "수정할 필드만 포함 (최소 1개 필수)",
        "minProperties": 1,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "할일 제목 (1~100자)"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true,
            "description": "할일 상세 내용 (최대 1000자, null 허용)"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "시작일 (YYYY-MM-DD)"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "description": "종료일 (YYYY-MM-DD, start_date 이상)"
          }
        }
      },
      "CompleteTodoRequest": {
        "type": "object",
        "required": ["is_success"],
        "properties": {
          "is_success": {
            "type": "boolean",
            "description": "완료 성공 여부 (true: 성공, false: 실패)"
          }
        }
      },
      "TodoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "todo": {
            "$ref": "#/components/schemas/Todo"
          }
        },
        "required": ["success", "todo"]
      },
      "TodoListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "todos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Todo"
            }
          },
          "total": {
            "type": "integer",
            "description": "필터 조건에 해당하는 전체 항목 수"
          },
          "page": {
            "type": "integer",
            "description": "현재 페이지 번호"
          },
          "limit": {
            "type": "integer",
            "description": "페이지당 항목 수"
          }
        },
        "required": ["success", "todos", "total", "page", "limit"]
      },
      "CompleteTodoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "todo": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "title": {
                "type": "string"
              },
              "is_completed": {
                "type": "boolean",
                "example": true
              },
              "is_success": {
                "type": "boolean",
                "description": "true: 성공, false: 실패"
              },
              "status": {
                "$ref": "#/components/schemas/TodoStatus"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": ["id", "title", "is_completed", "is_success", "status", "updated_at"]
          }
        },
        "required": ["success", "todo"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "오류 설명 메시지"
          }
        },
        "required": ["success", "message"]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "인증 실패 (Access Token 없음, 만료, 위변조)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "인증이 필요합니다"
            }
          }
        }
      },
      "Forbidden": {
        "description": "본인 소유가 아닌 리소스에 접근",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "해당 리소스에 접근할 권한이 없습니다"
            }
          }
        }
      },
      "NotFound": {
        "description": "할일을 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "할일을 찾을 수 없습니다"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "서버 내부 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "서버 오류가 발생했습니다"
            }
          }
        }
      }
    }
  }
}